Rename `preprocessor` to `syntax_rewriter`
