+++ crates/rome_js_formatter/src/syntax_rewriter.rs
